rxを使ってデータ更新時に通知を受け取れるように修正
