+++ Study-Himotoki-Realm-RxSwift/Base.lproj/Main.storyboard
@@ -35,7 +35,7 @@
                     <navigationItem key="navigationItem" title="Root View Controller" id="59U-Dv-yP7">
                         <barButtonItem key="rightBarButtonItem" title="Item" id="a0h-Q5-uNc">
                             <connections>
-                                <action selector="add:" destination="Sqk-Q0-yoY" id="MRb-PM-n8H"/>
+                                <action selector="didTouchUpItemButton:" destination="Sqk-Q0-yoY" id="L4M-DZ-u5P"/>
                             </connections>
                         </barButtonItem>
                     </navigationItem>
